requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-livereload-server==0.3.2
djangorestframework==3.12.4
idna==2.10
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
Pillow==8.3.1
pycparser==2.20
pydot==1.4.2
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.6
